moved clickable raycasting to own system
