+++ nc/source/physics/CMakeLists.txt
@@ -1,5 +1,6 @@
 target_sources(${ENGINE_LIB}
     PRIVATE
+        ${ROOT_DIR}/source/physics/ClickableSystem.cpp
         ${ROOT_DIR}/source/physics/ColliderSystem.cpp
         ${ROOT_DIR}/source/physics/ColliderTree.cpp
         ${ROOT_DIR}/source/physics/CollisionSystem.cpp
