+++ nc/source/physics/PhysicsSystem.h
@@ -1,20 +1,11 @@
 #pragma once
 
 #include "CollisionSystem.h"
-#include "physics/IClickable.h"
-#include "directx/math/DirectXMath.h"
-
-#include <vector>
+#include "ClickableSystem.h"
 
 namespace nc
 {
-    namespace graphics
-    {
-        class FrameManager;
-        class Graphics;
-        class Graphics2;
-    }
-
+    namespace graphics { class FrameManager; }
     namespace job { class JobSystem; }
 }
 
@@ -38,15 +29,8 @@ namespace nc::physics
     class PhysicsSystem
     {
         public:
-            #ifdef USE_VULKAN
-            PhysicsSystem(graphics::Graphics2* graphics2);
-            #else
             PhysicsSystem(const PhysicsSystemInfo& info);
-            #endif
 
-            void RegisterClickable(IClickable* toAdd);
-            void UnregisterClickable(IClickable* toRemove) noexcept;
-            IClickable* RaycastToClickables(LayerMask mask);
             void DoPhysicsStep();
             void ClearState();
 
@@ -56,11 +40,6 @@ namespace nc::physics
 
         private:
             CollisionSystem m_collisionSystem;
-            std::vector<nc::physics::IClickable*> m_clickableComponents;
-            #ifdef USE_VULKAN
-            graphics::Graphics2* m_graphics2;
-            #else
-            graphics::Graphics* m_graphics;
-            #endif
+            ClickableSystem m_clickableSystem;
     };
-} //end namespace nc::physics
+} // namespace nc::physics
